backend/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flasgger==0.9.7.1
boto3==1.34.108
python-dotenv==1.0.1
Flask-Cors
pandas

uuid
Pydantic # 推薦，用於類型驗證